requirements.txt
streamlit
dashscope
python-dotenv
Pillow